+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -180,7 +180,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfwkz5zpqee7j0hpqqwygxfs"),
+  _id: z.string().default("conv_01kfwv1x05evfzsjq0pg78pxg1"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),