+++ packages/orq-rc/RELEASES.md
@@ -6179,3 +6179,13 @@ Based on:
 - [typescript v4.3.0-rc.1] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.1 - packages/orq-rc
+
+## 2026-01-26 09:47:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.697.5 (2.799.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.2] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.2 - packages/orq-rc